Use the client-go exec plugin code
Instead of parsing the kube config ourselves and invoking a credentials
plugin if configured, we can use the client-go similarly to how we used
it for auth-provider users.

The catch is that the exec authenticator sets details on a transport
config object, but it is still possible to extract certificate
information and to use the request wrapping.

This feels slightly more hacky than the auth-provider code. But then
again, the k8s folks seem careful with their private methods, so we
might assume that the public interface will be maintained.

Issue: https://github.com/cloudfoundry/cf-crd-explorations/issues/64