Detect CF-on-K8s in `cf api`
This makes `cf api` able to detect the `cf-on-k8s` flag in the root
endpoint response and persist this information in the config. This will
allow us to enable CF-on-K8s-specific behaviour for every following
command.

We have introduced a new `selfcontained` integration suite which
contains tests that do not need to be run against a cf deployment.

Issue: cloudfoundry/cf-k8s-api#10
Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>