Implement cf login for CF-on-K8s
This implements a CF-on-K8s-compatible version of the `cf login`
command. When a user runs `cf login` against a CF-on-K8s, they will be
prompted to choose an `auth-info` from the ones specified in their
`$KUBECONFIG`. Passing any extra option, except for `-a`, will result in
a failure.

The main changes here are the extraction of the `AuthActor` and
`UserConfig` interfaces, which we implement both for the default use
case and the Kubernetes one. This allow us to provide specific
implementations of methods like `Authenticate`, `GetLoginPrompts` and
`CurrentUser*`.

While the choice of the `AuthActor` implementation can happen in the
`NewActor` constructor, this can't be done for `UserConfig`. `cf login`
hits the root endpoint and mutates the `Config` accordingly, and the
selection of the `UserConfig` implementation depends on those mutations.
We have then decided to perform this choice dynamically via the
`DynamicUserConfig` decorator.

Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>
Co-authored-by: Mario Nitchev <marionitchev@gmail.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>