Support inline and filepath tokens
Support the case where a token is provided directly in the auth-info
with `token` or using `token-file`.

The logic here is that there won't any certificates in the TLSConfig,
and the WrapTransport method of transportConfig will be nil, as it is
only set in the exec and auth-provider plugins. Therefore we'll just
wrap the http request with whatever auth info is left, which is
'hard-coded' tokens in this case.

Issue: https://github.com/cloudfoundry/cf-k8s-api/issues/11
Issue: https://github.com/cloudfoundry/cf-k8s-api/issues/12